add integration with PetFollow
Added helper class that a) check for the existence of the PetFollow mod in the current game, and if so, b) provides access to the core functionality of PetFollow through reflection.

The entire thing is wrapper in try/catch blocks to catch failures early and informatively, and hopefully not to spam errors on failure.